Move map stuff into the map package
